refactor(lock): move RepoLock logic to separate package
