+++ internal/lock/lock_test.go
@@ -1,4 +1,4 @@
-package main
+package lock
 
 import (
 	"strconv"
